LPS-129573 move to util package
